towards shorting the kernel(s) output

+++ library_daq.h
@@ -70,7 +70,7 @@ texture<unsigned int, 1, cudaReadModeElementType> tex_times;
 unsigned int * host_time_bin_of_hit;
 unsigned int * device_time_bin_of_hit;
 // npmts per time bin
-unsigned int * device_n_pmts_per_time_bin; // number of active pmts in a time bin
+unsigned short * device_n_pmts_per_time_bin; // number of active pmts in a time bin
 unsigned int * host_n_pmts_per_time_bin;
 unsigned int * device_n_pmts_nhits; // number of active pmts
 unsigned int * host_n_pmts_nhits;
@@ -104,8 +104,8 @@ cudaEvent_t start, stop, total_start, total_stop;
 bool output_txt;
 unsigned int correct_mode;
 // find candidates
-unsigned int * host_max_number_of_pmts_in_time_bin;
-unsigned int * device_max_number_of_pmts_in_time_bin;
+unsigned short * host_max_number_of_pmts_in_time_bin;
+unsigned short * device_max_number_of_pmts_in_time_bin;
 unsigned int *  host_vertex_with_max_n_pmts;
 unsigned int *  device_vertex_with_max_n_pmts;
 // gpu properties
@@ -134,7 +134,7 @@ unsigned int nhits_window;
 int n_events;
 
 
-__global__ void kernel_find_vertex_with_max_npmts_in_timebin(unsigned int * np, unsigned int * mnp, unsigned int * vmnp);
+__global__ void kernel_find_vertex_with_max_npmts_in_timebin(unsigned short * np, unsigned int * mnp, unsigned int * vmnp);
 __global__ void kernel_find_vertex_with_max_npmts_in_timebin_and_directionbin(unsigned int * np, unsigned int * mnp, unsigned int * vmnp);
 
 unsigned int read_number_of_input_hits();
@@ -1425,7 +1425,7 @@ void print_gpu_properties(){
 }
 
 
-__global__ void kernel_find_vertex_with_max_npmts_in_timebin(unsigned int * np, unsigned int * mnp, unsigned int * vmnp){
+__global__ void kernel_find_vertex_with_max_npmts_in_timebin(unsigned short * np, unsigned int * mnp, unsigned int * vmnp){
 
 
   // get unique id for each thread in each block == time bin
@@ -1580,7 +1580,7 @@ void copy_candidates_from_device_to_host(){
 
   checkCudaErrors(cudaMemcpy(host_max_number_of_pmts_in_time_bin,
 			     device_max_number_of_pmts_in_time_bin,
-			     n_time_bins*sizeof(unsigned int),
+			     n_time_bins*sizeof(unsigned short),
 			     cudaMemcpyDeviceToHost));
   checkCudaErrors(cudaMemcpy(host_vertex_with_max_n_pmts,
 			     device_vertex_with_max_n_pmts,
